… src/System.Text.Encodings.Web/src/System/Text/Encodings/Web/JavaScriptEncoderHelper.cs
… src/System.Text.Encodings.Web/src/System/Text/Encodings/Web/Sse2Helper.cs
… src/System.Text.Encodings.Web/src/System/Text/Encodings/Web/TextEncoder.cs
… src/System.Text.Encodings.Web/src/System/Text/Encodings/Web/UnsafeRelaxedJavaScriptEncoder.cs
… src/System.Text.Encodings.Web/src/System/Text/Internal/AllowedCharactersBitmap.cs
… src/System.Text.Encodings.Web/tests/JavaScriptStringEncoderTests.cs