Follow-up to https://github.com/dotnet/corefx/pull/41845 to optimize the built-in `JavascriptEncoder` using a similar approach.

Also added some tests to improve code coverage.

- Escaping UTF-16 (for strings of length >= 16)
   - using Default improved 2x
   - using Relaxed improved 3x
   - using Custom improved 15-20%
   - small regression for lengths <= 4.
- Escaping UTF-8 (for strings of length >= 16)
   - using Default improved 2-5x
   - using Relaxed improved 6-10x
   - using Custom improved 2-3x
   - small regression for lengths <= 2.

```C#
public unsafe class NeedsEscapingTest
{
    [Params(E.Default, E.Relaxed, E.Custom)]
    public E Encoder { get; set; }

    [Params(1, 2, 4, 7, 8, 15, 16, 17, 31, 32, 33, 47, 100, 1000)]
    public int DataLength { get; set; }

    public enum E
    {
        Default,
        Relaxed,
        Custom
    }

    private string _source;
    private byte[] _sourceUtf8;
    private JavaScriptEncoder _encoder;

    [GlobalSetup]
    public void Setup()
    {
        var random = new Random(42);

        var array = new char[DataLength];
        for (int i = 0; i < DataLength; i++)
        {
            array[i] = (char)random.Next(97, 123);
        }
        _source = new string(array);
        _sourceUtf8 = Encoding.UTF8.GetBytes(_source);

        _encoder = null;
        switch (Encoder)
        {
            case E.Default:
                _encoder = JavaScriptEncoder.Default;
                break;
            case E.Relaxed:
                _encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping;
                break;
            case E.Custom:
                _encoder = JavaScriptEncoder.Create(new TextEncoderSettings(UnicodeRanges.All));
                break;
        }
        _encoder.FindFirstCharacterToEncodeUtf8(_sourceUtf8);
    }

    [BenchmarkCategory(Categories.CoreFX, Categories.JSON)]
    [Benchmark (Baseline = true)]
    public int NeedsEscapingUtf16()
    {
        fixed (char* ptr = _source)
        {
            return _encoder.FindFirstCharacterToEncode(ptr, _source.Length);
        }
    }

    [BenchmarkCategory(Categories.CoreFX, Categories.JSON)]
    [Benchmark]
    public int NeedsEscapingUtf8()
    {
        return _encoder.FindFirstCharacterToEncodeUtf8(_sourceUtf8);
    }
}
```

Method | Encoder | DataLength | Before (ns) | After (ns) | Ratio
-- | -- | -- | -- | -- | --
NeedsEscapingUtf16 | Default | 1 | 2.845 | 5.359 | 1.884
NeedsEscapingUtf8 | Default | 1 | 5.286 | 5.32 | 1.006
NeedsEscapingUtf16 | Default | 2 | 3.836 | 5.651 | 1.473
NeedsEscapingUtf8 | Default | 2 | 6.866 | 6.117 | 0.891
NeedsEscapingUtf16 | Default | 4 | 5.96 | 7.406 | 1.243
NeedsEscapingUtf8 | Default | 4 | 11.188 | 7.489 | 0.669
NeedsEscapingUtf16 | Default | 7 | 9.746 | 10.047 | 1.031
NeedsEscapingUtf8 | Default | 7 | 18.271 | 9.449 | 0.517
NeedsEscapingUtf16 | Default | 8 | 10.732 | 8.708 | 0.811
NeedsEscapingUtf8 | Default | 8 | 19.119 | 9.904 | 0.518
NeedsEscapingUtf16 | Default | 15 | 18.068 | 14.783 | 0.818
NeedsEscapingUtf8 | Default | 15 | 35.961 | 14.552 | 0.405
NeedsEscapingUtf16 | Default | 16 | 19.266 | 10.416 | 0.541
NeedsEscapingUtf8 | Default | 16 | 34.358 | 9.519 | 0.277
NeedsEscapingUtf16 | Default | 17 | 20.162 | 11.761 | 0.583
NeedsEscapingUtf8 | Default | 17 | 38.229 | 9.16 | 0.240
NeedsEscapingUtf16 | Default | 31 | 35.119 | 22.04 | 0.628
NeedsEscapingUtf8 | Default | 31 | 70.864 | 18.698 | 0.264
NeedsEscapingUtf16 | Default | 32 | 37.013 | 18.846 | 0.509
NeedsEscapingUtf8 | Default | 32 | 80.238 | 10.797 | 0.135
NeedsEscapingUtf16 | Default | 33 | 41.737 | 20.024 | 0.480
NeedsEscapingUtf8 | Default | 33 | 74.927 | 11.542 | 0.154
NeedsEscapingUtf16 | Default | 47 | 53.155 | 31.701 | 0.596
NeedsEscapingUtf8 | Default | 47 | 102.116 | 20.617 | 0.202
NeedsEscapingUtf16 | Default | 100 | 118.166 | 57.782 | 0.489
NeedsEscapingUtf8 | Default | 100 | 208.161 | 23.387 | 0.112
NeedsEscapingUtf16 | Default | 1000 | 1,068.02 | 537.261 | 0.503
NeedsEscapingUtf8 | Default | 1000 | 2,077.12 | 129.447 | 0.062
  |   |   |   |   |  
NeedsEscapingUtf16 | Relaxed | 1 | 3.523 | 4.45 | 1.263
NeedsEscapingUtf8 | Relaxed | 1 | 5.164 | 5.382 | 1.042
NeedsEscapingUtf16 | Relaxed | 2 | 6.617 | 5.444 | 0.823
NeedsEscapingUtf8 | Relaxed | 2 | 9.37 | 6.744 | 0.720
NeedsEscapingUtf16 | Relaxed | 4 | 7.057 | 7.594 | 1.076
NeedsEscapingUtf8 | Relaxed | 4 | 13.711 | 9.583 | 0.699
NeedsEscapingUtf16 | Relaxed | 7 | 9.742 | 10.223 | 1.049
NeedsEscapingUtf8 | Relaxed | 7 | 20.175 | 13.549 | 0.672
NeedsEscapingUtf16 | Relaxed | 8 | 10.666 | 5.822 | 0.546
NeedsEscapingUtf8 | Relaxed | 8 | 20.463 | 15.222 | 0.744
NeedsEscapingUtf16 | Relaxed | 15 | 18.347 | 13.148 | 0.717
NeedsEscapingUtf8 | Relaxed | 15 | 36.851 | 24.855 | 0.674
NeedsEscapingUtf16 | Relaxed | 16 | 19.439 | 7.579 | 0.390
NeedsEscapingUtf8 | Relaxed | 16 | 39.43 | 6.166 | 0.156
NeedsEscapingUtf16 | Relaxed | 17 | 19.919 | 8.377 | 0.421
NeedsEscapingUtf8 | Relaxed | 17 | 42.782 | 7.535 | 0.176
NeedsEscapingUtf16 | Relaxed | 31 | 34.941 | 18.058 | 0.517
NeedsEscapingUtf8 | Relaxed | 31 | 78.846 | 27.82 | 0.353
NeedsEscapingUtf16 | Relaxed | 32 | 36.679 | 13.355 | 0.364
NeedsEscapingUtf8 | Relaxed | 32 | 80.051 | 8.601 | 0.107
NeedsEscapingUtf16 | Relaxed | 33 | 38.289 | 14.08 | 0.368
NeedsEscapingUtf8 | Relaxed | 33 | 83.212 | 10.119 | 0.122
NeedsEscapingUtf16 | Relaxed | 47 | 52.697 | 23.491 | 0.446
NeedsEscapingUtf8 | Relaxed | 47 | 112.786 | 33.978 | 0.301
NeedsEscapingUtf16 | Relaxed | 100 | 119.245 | 41.575 | 0.349
NeedsEscapingUtf8 | Relaxed | 100 | 235.103 | 21.904 | 0.093
NeedsEscapingUtf16 | Relaxed | 1000 | 1,083.64 | 360.613 | 0.333
NeedsEscapingUtf8 | Relaxed | 1000 | 2,306.57 | 142.98 | 0.062
  |   |   |   |   |  
NeedsEscapingUtf16 | Custom | 1 | 2.794 | 3.789 | 1.356
NeedsEscapingUtf8 | Custom | 1 | 4.726 | 5.637 | 1.193
NeedsEscapingUtf16 | Custom | 2 | 3.679 | 4.455 | 1.211
NeedsEscapingUtf8 | Custom | 2 | 7.586 | 6.657 | 0.878
NeedsEscapingUtf16 | Custom | 4 | 5.993 | 6.336 | 1.057
NeedsEscapingUtf8 | Custom | 4 | 12.065 | 8.793 | 0.729
NeedsEscapingUtf16 | Custom | 7 | 9.794 | 9.196 | 0.939
NeedsEscapingUtf8 | Custom | 7 | 20.017 | 12.328 | 0.616
NeedsEscapingUtf16 | Custom | 8 | 10.898 | 9.973 | 0.915
NeedsEscapingUtf8 | Custom | 8 | 21.423 | 13.255 | 0.619
NeedsEscapingUtf16 | Custom | 15 | 18.924 | 16.201 | 0.856
NeedsEscapingUtf8 | Custom | 15 | 37.769 | 21.461 | 0.568
NeedsEscapingUtf16 | Custom | 16 | 21.059 | 17.102 | 0.812
NeedsEscapingUtf8 | Custom | 16 | 39.721 | 17.138 | 0.431
NeedsEscapingUtf16 | Custom | 17 | 21.333 | 17.613 | 0.826
NeedsEscapingUtf8 | Custom | 17 | 42.232 | 18.459 | 0.437
NeedsEscapingUtf16 | Custom | 31 | 35.81 | 30.757 | 0.859
NeedsEscapingUtf8 | Custom | 31 | 77.807 | 37.533 | 0.482
NeedsEscapingUtf16 | Custom | 32 | 37.572 | 31.069 | 0.827
NeedsEscapingUtf8 | Custom | 32 | 81.055 | 28.39 | 0.350
NeedsEscapingUtf16 | Custom | 33 | 37.351 | 30.971 | 0.829
NeedsEscapingUtf8 | Custom | 33 | 85.248 | 30.07 | 0.353
NeedsEscapingUtf16 | Custom | 47 | 54.932 | 43.93 | 0.800
NeedsEscapingUtf8 | Custom | 47 | 115.287 | 48.567 | 0.421
NeedsEscapingUtf16 | Custom | 100 | 118.678 | 91.347 | 0.770
NeedsEscapingUtf8 | Custom | 100 | 237.319 | 79.042 | 0.333
NeedsEscapingUtf16 | Custom | 1000 | 1,066.93 | 882.03 | 0.827
NeedsEscapingUtf8 | Custom | 1000 | 2,288.41 | 722.101 | 0.316



## Next Steps:
1) Evaluate if we can get rid of the special-casing in `S.T.Json`

cc @GrabYourPitchforks, @gfoidl, @tarekgh, @tannergooding, @ViktorHofer, @steveharter